Add kernel support for debugging with gdb. This also make /proc allow access to the registers. The format of the registers returned in ptrace, /proc and in the coredump header match what gdb was expecting. Floating point registers are not yet copied into the structure. The location of the fpr needs to be found and copied. <- TODO
Changes to trap.c was to add some interm debugging. code is commented out.

powerpc does not currently have a MID_* type, since it uses ELF instead
of a.out, one was never added. Currently the coredump is done in form
of MID_NONE. What is the correct approach for this?? Elf format coredumps...
maybe?