Bug fixes from NetBSD, port-powerpc/7240 and port-powerpc/7243
The first fixes a major problem seen _too_ many times to record.
If a task that is busy doing disk io is interrupted with a ^C,
its would frequently put the system into a mode where it would
very shortly crash, It was known that it was a signal delivery
that was occurring incorrectly, but had not been diagnosed where.
It is no longer possible to crash the machine on whim.
(This probably should be recorded as a stability patch for 25).

The second was not a bug that had ever been seen to be identified,
but looking at the code it was obvious that the onfault variable was not
being reset before the function exit. The code differs between NetBSD
and OpenBSD somewhat, however only two instances of the code was
reported, and three instances exist in the OpenBSD code. (all were fixed).