Add "expert mode" and don't prompt for fsize/bsize/cpg unless an expert
next_offset() now takes a pointer to the partition size as its second
argument instead of a pointer to the partition.  This allows us to
get the offset earlier on and sort_partitions() will not try and
include the partition we are trying to create in its list.

In free_chunks() start looking for things at starting_sector, not 0,
to avoid getting starting offsets that precede the start of the OpenBSD
part of the disk.