From FreeBSD:
Yet another bug fix/optimization for the Davicom DM9100/9102: increase the
PCI latency timer value to 0x80. Davicom's Linux driver does this, and it
drastically reduces the number of TX underruns in my tests. (Note: this is
done only for the Davicom chips. I'm not sure it's a good idea to do it
for all of them.)