Deal with files that lack a final newline before EOF (you naughty emacs users!).  In most cases this just means checking for EOF in addition to '\n'.  However, we need to tread carefully in readhash(). There we return 0 on EOF iff it is the first char in a "line".
Additionally, if the -b or -w flags were specified and the last
character in one file was '\n' but the other file is missing '\n',
pretend that we didn't see the newline in the one file.  This is
consistent with GNU diff.  For the non-{b,w} case, print
"\ No newline at end of file" when we see EOF before a newline in
one file where the other file has one.  Again, this is for consistency
with GNU diff and patch(1) in -current knows how to interpret this
message.

OK tedu@ and otto@