Caps Lock key is special, even in raw mode, on the Domain keyboard, exactly like on ADB keyboards. Must have been the 80's keyboard fashion.
So when we see a Caps Lock event, produce the missing event as well for
acceptable behaviour.

This does not allow the Caps Lock key to be used for chording, though,
so document this in the manpage (honestly, Ctrl and Caps Lock being adjacent,
there is no point in swapping them on Domain keyboards, but you never know).