ami chips have no mechanism to take commands off them once we've submitted them. this means that we cant reliably complete an io before the chip says we can because it can possibly complete later and overwrite memory it no longer owns, or write garbage to disk.
so this diff forces the timeout on a scsi io to be as long as the chip
should ever take, which is sixty seconds. after much discussion with marco
we also decided to add a bit more so the time spent in the software runq
wouldnt affect the ios lifetime too much.

instead of completing ios out of the timeout, this simply warns the
operator to check the state of the volumes if things are starting to slow
down.

ok marco@