add the following paragraph from the wi(4) manual page to the CAVEATS section of all drivers claiming HostAP support:
"Host AP mode doesn't support power saving.  Clients attempting to
 use power saving mode may experience significant packet loss
 (disabling power saving on the client will fix this)."

unfortunately (or not?), more and more clients are using PS mode by
default and will not work properly with an AP running OpenBSD.

proper handling of STAs in PS mode will come in net80211 but is is
very unlikely that all drivers will be ported.