We don't count buffercache stats in the B_PHYS case, so fix nfs to not increment the num{read,write} and pending{read,write} statistics in that case, since biodone won't change them on completion.
On another note, I'm not sure that we use physical buffers for swapping
over nfs anymore, so this chunk may be superfluous.

beck@ came up with the same diff "So anyway rather than me commiting it
from my copy, I'm giving you the OK and the commit. since it officially
makes you a buffer cache and NFS hacker };-)"