Make .de1 a symnonym for .de, not .ig as it was before.
The .de1 instructions is a GNU extension not found in traditional roff and
not even in old groff, defined as "define a macro that will be executed with
traditional roff compatibility mode switched off during macro execution".
Since we ran into it in the wild, we have been parsing and ignoring it for
a long time.  Now that we have proper .de support, we can as well use the
contents, even though we don't implement compatibility mode at all.