Disable the 88110 branch prediction logic unconditionnaly on all 88110 revisions; despite what the ``official'' (yet unpublished, confidential proprietary, will cause a tree to fall on your house if you quote it, etc) errata says, disabling data decoupling is not enough to workaround its malfunction in processor revisions 5.x.
Enough missing-SFU instructions (each causing a `disabled SFU' trap) in a
tight loop will eventually (but quickly) trigger the (unrecoverable, not even
by NMI) processor hang.

Of course, most such instructions are not privileged, and can be easily issued
by an evil userland process; crashme happens to be a good example of this, when
invoked with the proper settings (which are left as an exercise to the reader).

Now, can I have my hair back? Come on! Please... pretty please... with sugar on
top... people are looking at my head, you know.