provide the exchange rate for those currencies listed in the federal reserve section (i'll get to the others sometime), as of 23.9.11. i've also updated the website link, and the date of the exchange rate.
because the rates given on the federal reserve website are in currency
units per US dollar, and units(1) needs US dollars per currency unit,
there is an additional layer of conversion and rounding (and work) to
perform. i have side stepped that by prefixing all the amounts by "1|"
to let units(1) do the work. for whatever reason, four currencies are
given in US dollars per currency unit and do not need this.