Make xlint go the way of the Dodo.
Lint had a great run but it is unmaintained and there are much better
tools out there these days.  The output has too many false positives and
therefore is completely ignored by developers.  It pains me to do this
but it is time to admit that lint as is dead.