Retire a couple of filesystems.
procfs: It's i386 and for linux compat support only; The linux compat
support is also on it's way to the chopping block.

ntfs: This is a crippled (read only) and unmaintained port that noone
has ever cared about.

nnpfs/arla/afs: Been disabled for a long time, and not worth continuing.
There is however a horrible hack in kerberos now to keep it compiling.