Move the uvm_pagealloc reserve handling to pmemrange.
This way all allocations get the benefit so they can't starve important
processes when we are close to memory starvation.

This adds a usereserve flag to the pmemrange/pglistalloc flags.

The semantics also change slightly. We no longer give reserve pages to
allocations into the kernel object, leaving reserve for pmap (and other
uses of the usereserve flag) and the pagedaemon. Finally, fix a bug
where we could deny the pagedaemon memory when there was stil left by
implicitly counting the pagedaemon has being allowed the reserve,
previousy the check was backwards.

Since pmemrange allocates multiple pages the check has to be for if (free <=
limit + sizeof(allocation)) now.

much discussion with ariane@