Remove spurious calls to pmap
This commit is the result of an audit of callers of
pmap_clear_reference() and pmap_page_protect(,,VM_PROT_NONE)

Remove a bunch of uneeded calls to pmap_clear_reference():
In uvm_anon.c and uvm_aobj.c it is only called on swapoff. We don't

really require the reference cleared if it has been accessed. doesn't
matter. and if it was just swapped then the page won't be deactivated
and it'll be freshly cleared again. However, it is possible that this
code is still wrong: deactivating something we just swapped in is wrong,
since it isn't swapped out and has no backing store. a deactivated page
mean we can dump them at will.

With pmap_page_protect, every caller of uvm_pagedeactivate first calls
pmap_page_protect. This is techinically incorrect, but done due to pmap
bugs.  move the page protect call into uvm_pagedeactivate itself and
detail with an XXX comment why this is wrong.