Make the interrupt handler chains use a real lock instead of si_slock
The locking protocol was entirely broken here, fix it up. Now we just hold a
spinlock while we have that interrupt masked (we preallocate the descriptor) to
prevent other cpus jumping in and screwing with the chains while we do so. We
have masked anything that we are touching so we won't be interrupted either.

Remove `locking' on idt vectors since they are only ever touched from
intr_{dis,}establish().

A couple of nits, some cleanup and ok from pedro (vote for pedro!).