Avoid a pagedaemon deadlock (and panic).
A uao that uses the swhash inserts a deadlock if we do a swslot_alloc
during the pagedaemon (where by necessity we have the pageqlock and the
object locked) to allocate a hash bucket.

A proper fix has been designed and is being worked on, but for now just
make the pool_get NOWAIT. Since this is always in the pagedaemon you
will be *always* calling this where we can use the pagedaemon reserve
for allocations. If an allocation fails in this case we may as well just
panic because we are *fucked*.

discussed with and semi-instigated by ariane@.