Retire disksort.
Arguably, all IO sorting is bad (as it primarily adds latency to the IO)
and even if it could help, disksort is always slower then a straight up
FIFO in all cases I've tested. The disk will do a better job of it.

Clean up a bunch of now unused members (defines) in struct buf.

Makes marco all warm and fuzzy inside.