Wrap rtable_add() and rtable_l2set() calls inside ifioctl() inside softnet, as functions that modify routing information shouldn't be interruptable by network traffic.
Also make sure that both of those functions assert that they are called
at softnet. I'm reasonably sure that there shouldn't be any, but if there
is any codepath that was missed, we're going to be here for another 4 days
to deal with any fallout.

While here, move the multitude of "int s" declarations inside ifioctl to
the beginning of the function.

okay claudio@