Rework crtbegin code.
Instead of adding a -1 at the beginning of list, just make the ?TORS_LIST
an empty array like JCR_LIST. This allows the skipping of the first
element and the code to be cleaned up and simplified.

Move some elements to __dso_hidden and not static/const so that clang
will not try to optimize away the needed functions/variables.

Commonize the code between crtbegin and crtbeginS.c as much as possible.