fix KMS on ironlake (gen5) gpus.
After mapping the pipecontrol buffer and the status page, wire both
mappings immediately. The bug is the following:

with the pc buffer, we map it but don't fault it in until the first time
we need to check a seqno, which we do with a lock held.
we check the buffer, we fault and try to grab the map lock.
rw_enter_read -> assertwaitok -> mutex held, boom

The same bug occurs with the status page but we never hit it since we
memset it after map, faulting it in. Though if we ever try and swap it
out we could hit bad things.  Bug also exists on openbsd but they don't
have assertwaitok in rw_locks so don't hit the assert before anything
bad happens.  Wiring the page avoids swapping it and faults it in
avoiding all problems.