Prevent the need to build twice when libc major bumps
For proper TLS operation, it is necessary for libraries to link against
libc, however if libc major bumps and is built (without installing it)
before the other libraries are built, make build will cause libraries
to depend on a different version of libc than the binaries, thus bringing
both old and new libc into the executables. This fixes the issue
by building and installing libc before building the rest of the libraries,
thus the libraries will link against the same version the binaries use.