Fix the locking for uao_grow and uao_shrink.
These functions want to be able to sleep for memory. However in order to
be safe they need to lock the aobj correctly. There is a race where one
owner of the aobj could resize it while the other was sleeping for
memory (or vice versa) and this could change whehter it was a grow or a
shrink (this really would be a case where the aobj was shared by two
disparate owners that do not synchronise, but this does not mean that it
should not have defined semantics instead of hitting an assert). for
that reason, replace uao_{grow,shrink} with uao_setsize which handles
these cases.

v2: handle case where uvm_flush can sleep too, reminded by pedro.

ok pedro@