Fix deadlock in ufs_rename() caused by WAPBL
We cannot allow vrele() or vput() to cause a flush of the WAPBL journal
halfway through a rename operation, since the rename operation is a log
transaction itself. Failure to observe this would cause a hang where we
would try to grab a write lock (to flush the journal) even though we
held a read lock (to prevent the journal from being flushed).

The solution is to have ufs_rename() wrapped, grabbing an extra
reference to every vnode involved in the rename, and to have them all
disposed of only after the rename is complete.

OK marco@