Make 'make -j' respect .ORDER directive.
When checking if a .ORDER directive had it predecessor completed,
do not check the pred state of UNKNOWN, because it could be building
or several other states. Instead have has_unmade_predecessor() check
that the predecssor's state is not (MADE, UPTODATE, or ERROR).

Previously the predecessor state would change from UNKNOWN to BEINGMADE,
which allowed the successor object to start building before the predecessor
completed.

Diff sent to espie@openbsd.org and ok oga@