Adjust ffs_wapbl_start() call in ffs_mountfs()
Move the call to ffs_wapbl_start() a bit further down in ffs_mountfs(),
after um_savedmaxfilesize has been initialised. This is necessary
because ffs_wapbl_start() triggers an update of the superblock through
ffs_sbupdate() and if um_savedmaxfilesize isn't set at that point,
the maxfilesize field in the superblock will be set to zero.

OK marco@