Fix spec_fsync() on devices mounted with WAPBL
When a FFS file system is unmounted, its device vnodes are reassigned to
specfs. These device vnodes may be mounted somewhere themselves, and it
then becomes spec_fsync()'s responsibility to sync them. It currently
does so by iterating through the vnode's dirty buffers and
asynchronously pushing them, much like vflushbuf() does. That is not
enough for WAPBL, however. WAPBL requires ffs_wapbl_fsync_vfs() to be
called in order for the journal to be correctly flushed.

OK marco@