The end LBA is inclusive, so we need to fix some calculations.
If start LBA equals end LBA, a partition has a size
of one sector.  This means, we can't have zero-sized
partitions, as end LBA would be smaller than start LBA.

The size of a disk is the first unusable sector, which means
no LBA should be equal to that size.