replace broken tmpfs_dircookie() logic
the tmpfs_dircookie() logic keeps state of the last node returned by
VOP_READDIR() on a given directory and gives hints as to the offset
of other nodes to accelerate traversal. the logic is implemented by
taking the address of the leaf node, truncating it to 31 bits, and
using the result as the cookie. this approach has two problems: a) it
leaks kernel pointers to userland; b) there is no longer a need for
the 31-bit truncation since linux compat is gone.

this change implements a smarter logic based on work by rmind@netbsd.
instead of the address-derived cookie, we use the complete 2^64 range
to assign each leaf node an unique sequence number, and this number
is then used as the cookie. this approach has the drawback that
exhaustion is possible after 2^64 allocations on a given directory.
such a situation, however, is extremely unlikely to ever happen. in
case it does, the new cookie logic will prohibit new nodes from being
added to the directory until the necessary removals take place.

ok oga@ thorduri@