Fix crash on urndis(4) detach
When detaching my nokia c2-01 dohooks in if_detach would panic.

This was due to a failure to setup a pipe in urndis_ctrl_init()
which results in a TIMEOUT on receive with side-effects later on
during detach due to assumptions regarding the existence of interface
hooks during dohooks on if_detach tear-down.

That resulted on dereferencing a NULL function pointer which triggered
a panic.

Discussed with mpi@, tested and okay giovanni@.