Delay building ldconfig and ldd until normal build time.
When ld.so was added to the early build list, this caused it to build ldd
and ldconfig early as well. This is not desireable as they then build
against the old headers/libraries. Do not build these binaries until
later by adding them to SKIPDIR.

NOTE: NOMAN is part of the sudo environment to preserve, however SKIPDIR
is not, so the <set env> \${SUDO} needs to be \${SUDO} <set env>

Remove exec from 'exec \${SUDO} <set env>'.

$ exec {EMPTYVARIABLE} FOO=bar cmd doesn't work, and the exec only
skips a single fork, so rather go thru utter pain to solve this,
remove the exec.

ok patrick@