Make bus_space_barrier on i386 actually work
Convert bus_space_barrier to an inline function that always provides
an instruction barrier to prevent compilers from possibly reordering
stores around it.

In the case of a read barrier, we also do a locked add to flush all
write combining buffers on the bus.  The locked add is provided
versus using *fence instructions as the latter requires sse2, which
not all i386 have.

from FreeBSD.

ok oga@