Fix race in tmpfs_vnode_get()
Set v_data before the call to checkalias(), as we might sleep there
after the vnode is inserted in the hash chain. This makes it possible
for another process to concurrently call checkalias(), find the vnode we
were originally allocating and try to vget() it, which needs v_data set
for locking purposes.

With input from and OK oga@