Revert "Fix block allocation path in WAPBL"
This reverts commit 38d8b286e5bcef22adb38f816b290cbcb5f36062.

WAPBL suffers from the fundamental problem that, unlike softdep, it does
not guarantee that newly allocated blocks will make it to disk before
the metadata pointing to them. The approach taken in 38d8b286e5bce fixes
this problem but imposes unacceptable performance penalties by
completely disrupting the write clustering code. Revert it for now while
a proper solution is worked on.