sprinkle const in ffs1_blkpref(), ffs2_blkpref()
make ffs1_blkpref() and ffs2_blkpref() receive a const pointer to the
array of block pointers to make it clear that they are not modified
within those functions.

while there, remove pointless assignment from ffs1_blkpref() and make
the pointer to the superblock a const in both functions.

OK oga@