Convert kernel lock to a rrwlock.
Introduce rrw_exit_all(), rrw_enter_cnt() and rw_held()/rrw_held().

Also, make the assert unlocked functions make sense, now we assert
that _we_ don't hold the lock, asserting that no one has the lock is
impossible.

We can't sleep on proc_trampoline_mp for the idle process, since it
should never be in a sleepqueue, I had to introduce a hack to make it
not acquire the kernel lock.

The correct fix is to pass a flag to fork1() that gets propagated to
proc_trampoline_mp().