When sending a neighbor advertisement, correctly use the carp mac address and not the mac of the physical interface as target lladdr.
This allows us to get rid of the hack adding and removing the route to
loopback of a carp node when its status switch to master and backup
respectively.  Now the route entry will no longer be updated to use the
mac address of the physical interface of the master node as gateway.

While here send the ARP request through the carp interface for consistency.

Analysed with and ok mikeb@, ok bluhm@