Fix real syncs in softraid
Perform a wakeup on the discipline whenever there is a sync in progress
and we are disposing of a workunit. This allows non-fake syncs (such as
the ones performed by WAPBL) to work on a softraid volume without timing
out. Debugged with much help from Martin Natano (natano@).

OK oga@