Remove unnecessary check for B_LOCKED
B_LOCKED means that the buffer was added to a WAPBL transaction and
therefore needs to remain locked until the journal is written. It is
then important to notify WAPBL whenever the size of a B_LOCKED buffer is
modified.

Such a modification happens in bread_cluster_callback(), however it is
not possible for B_LOCKED to be set on that particular buffer, since we
don't perform read-ahead on metadata.