Remove misleading comment
It turns out that the initial assessment of wl_bufcount_max was quite
reasonable, so remove the comment which claims otherwise. Verified
against NetBSD.

The rationale is to bind wl_bufcount_max by half the number of available
buffers, multipled by 1024 (taken as a very small fragment size). To
approximate this value, bufpages is multipled by the page size and
divided by a conservative estimate of 16K per buffer. (A more empirical
value would be around 4K, but due to the dynamic buffer cache we read in
whole 16K buffers.)