Newly formed processes should start in a critical section.
They have the schedlock, and release on proc_trampoline_mp(), if the
newly formed process is not in a critical section, a clock interrupt
might attempt to grab the SCHED LOCK, or worse, we might preempt to
early in the future.

This paves the way for preemption and allows SCHED LOCK to be a mutex,
we only need guenther's fix for the recursive tsleep.