Initial kernel preemption.
This diff introduces a rudimentary form of kernel preemption, it
allows interrupt threads to preempt other kernel processing, like
userland doing syscalls or other normal kthreads.

Preemption is deferred if the current process is in a critical section.

When curproc preempts:
  o It must not be stolen to run on another cpu, it
is still SRUNable, but must therefore be pinned to the curcpu()
runqueue.
  o It must continue to hold all the locks it had, in our case:
rwlocks, including kernel lock.
  o In case the _preempting_ process (not the preempted) tries to
acquire the kernel lock, but the holder was preempted, the
_preempting_ process blocks so that the _preempted_ might be resumed.

CAVEATS:

  o The way we test for preemption is still primitive, I'll only be
able to attack this properly once I introduce proper APIs into the
scheduler, but for that I want to make the scheduler modular, so that
we are able to play with different schedulers. So we need MORE APIs,
not LESS as _some_ people believe.
  o ci_want_resched which is responsible for preempting userland is
still present as at this point we need it, once I'm able to hack the
scheduler there should be only one way of preemption, so it will look
a bit funky until there.

Also introduce a sysctl (kern.preemption) which enables us to activate
or deactivate kernel preemption.