Fetch list of mirrors in the installer.
This requires SSL in ftp(1), so distrib/special/ftp was removed. The
RAMDISK had to grow to accomodate SSL and the full ftp(1).

Since we don't distribute sets in the cd, default the installer to
http in case we're booting from the cd.

NKL2_KIMG_ENTRIES has been bumped to accomodate RAMDISK growth as
well.

Thanks to drahn@ for helping lots with the distrib stuff.

ok pedro@