Move aoe to the attic
Move aoe to the attic. It was put on the tree to ease hacking (in the
not so golden years of CVS) but was left unfinished and in need of work,
so sweep it aside for now until it is developed on a separate branch and
ready to be merged back in.