Prevent ls/stat -F from printing double slashes.
The -F flag caused ls to print a slash after all directory names, even
the ones which already contain a traling slash. Change that to prevent
the output of odd looking double slashes. POSIX requres ls -F to "Write
a slash ('/') immediately after each pathname that is a directory[...]".
It can be argued, that in the case where the pathname already contains a
trailing slash, this requirement is already sufficiently satisfied.

ok pedro@