Put AF_ROUTE socket that arp(8) operates on into the appropriate rdomain.
Previously "arp -V 1 -d 10.0.0.1" would hang forever since while the
target rdomain was passed in the rtm, the AF_ROUTE socket was always
created in the rdomain 0.  Due to the logic in route_input (rtsock.c)
route socket's rdomain is compared against the target rdomain of the
request and reply is generated only when they match.  In our case
rdomains were different which cased arp(8) to hang waiting for the
reply that would never be sent.

ok deraadt