Fix a race in vget() on allocating and inserting inodes into the chain.
Under specific circumstances, it could happen that a Inode is allocated,
as it's not found in the inode hash chain
When an inode is searched for, and not found in the hash chain, a
vnode/inode pair is allocated and inserted into the chain.  Between
allocation and insertion, vget() might sleep.  During that time, another
process can allocate and insert the same inode.  When the original
process then tries to insert into the chain, the inode is already there.

Now, the vnode is released, which will cause the inode to be removed
from the chain.  As the inode has never been inserted into the chain,
this will crash.

To fix it, set a flag to tell the chain remover that the inode
is not in the chain.

Debugged, found and fixed by pedro@