Fix hash bucket optimisation in uao_shrink_hash()
Call uao_shrink_flush() when doing the hash buckets optimisation in
uao_shrink_hash(). Failure to do so means that pages are not flushed,
and the anonymous object left in an inconsistent state.

For the record (and since I tend to forget things five minutes after
writing them): the point of this optimisation is not performance per se,
but rather a logical avoidance of malloc() in a code path whose purpose
is to free memory.

Adjust some comments while here.

OK oga@