Fix hash bucket optimisation in uao_grow_hash()
The assumption made by uao_grow_hash() that the buckets for the pages
being copied should not change can be violated under rare circumstances,
if and only if the hash mask changes. Detect this case and resort to
iterating over the existing buckets, effectively removing the elements
from the old hash table and inserting them afresh.

OK oga@