Implement tmpfs snapshot support.
The future purpose of tmpfs snapshots is to be used in a special device driver
that allows us to use snapshots as a ramdisk.

Snapshots can be created like this:

    # mount -t tmpfs -u -o snapshot /path/to/snapshot /mnt

Note that the snapshot file must reside on a different mountpoint than the one
which is to be dumped.

They may be mounted like this:

    # mount -t tmpfs /path/to/snapshot /mnt

written by pedro@ and me

ok patrick@; eye-balled by oga@