Add support for the Broadcom PHY found on the Octeon-based DSR-500.
This is an MI driver currently targeting only the BCM53115 model,
but other Broadcom devices (specially from the 53XX family) can make use
of it as well.

The driver currently accounts just for the CPU port. The switch is left
in dumb-mode. Further advanced switch control is in the works.

Parts of this was inspired by looking at the b53 driver from the
OpenWrt project. Thanks!

Okay miod@