Allow dynamic uvm_page_physload() on amd64.
Allow uvm_page_physload() to be called after uvm_init() on amd64. This
never really worked on most architectures because there is no mechanism
to notify pmap about the new pages. However, if the physical mappings
are in place and the architecture is __HAVE_PMAP_DIRECT, this should
work.

Careful testing seems to confirm this hypothesis.

ok oga@ patrick@