Make mkuboot fill in the eramdisk variable.
The pointer to the end of the tmpfs ramdisk will be stored in
eramdisk.  To do this on arm, which does not have its boot(8)
to do that, we need to fill it in manually when creating the
u-boot image.

The eramdisk variable is directly after esym, which means that
we can use the current code to insert the symbol table pointer
and add code to add the ramdisk pointer right after it.

The tmpfs ramdisk has to be stored page aligned and padded
until page aligned.  This means that eramdisk also has
to point to the page aligned end.

This means we need to fill the gap between the symbol table
and the ramdisk, and to pad the ramdisk after it has been written.