Update FDT handling in arm machdep.
Due to the recent PIC startup code change, we do not
have access to the previous VA==PA mapping.

Previously parts of arm's machdep were based on the view
of having that space mapped.  Now, we need to make sure
we do not access unmapped areas, so we use the bootstrap
mappings to give us a little bit of access.

Map the bootconfig/FDT space, make sure we have mapped
all of it.