skip empty PT_BITRIG_TMPFS_RAMDISK segments on amd64
version 3.29 of the bootloader saw the addition of code to detect and
load PT_BITRIG_TMPFS_RAMDISK segments contained in the kernel image.
this code, however, did not check whether such segments were empty
before rounding up their size to PAGE_SIZE. with the introduction of
ldscripts for amd64 and arm, GENERIC kernels gained empty RAMDISK
stub segments for future use by rdconfig(8).

given that tmpfsrd is disabled in GENERIC,  this situation is not
problematic on its own. once it gets enabled, as is our plan to kill
RAMDISK kernels, tmpfsrd will detect and attach to these rounded-up
stub segments, a behaviour which, while harmless, is clearly unintended.

to address this situation, this change does:

- add the missing check for empty PT_BITRIG_TMPFS_RAMDISK segments to
  the bootloader.
- crank the bootloader version number.
- add a failback temporary check to tmpfsrd to check whether the size of
  the segment being attached to is larger than a page, in which case we
  don't attach. this is meant as a interim solution so 3.29 bootloaders
  can boot exactly the same kernels as 3.30, making transition smoother.

ok patrick@ natano@