final bulk shows chunking to happen in all the right places, and signing works correctly as well. So activate this.
This basically makes package snapshots rsync-friendly with a low-tech
approach. Instead of having a patched gzip, we borrow from other archive
formats, and cut the archive every few files. Since the files are already
sorted in order of date of modification, packages that don't change tend
to end with identical gzip archives.

Note that this will still work when/if we transition to bzip2/xz packages...