during a small refactor, we broke table_passwd.
lookup function rely on a parsing function to have a struct passwd members
point to the appropriate position in a buffer... but instead of passing the
buffer, the parsing function uses its own internal buffer which will no
longer be valid upon return.

turns out to work by luck in my tests, not so much when Daniel Adolfsson
tried to use it.

spotted by Daniel Adolfsson <daniel@priv.nu>