clang: make aarch64 target work again
In the ports clang, we had support for the AArch64 target.  When
we imported it into base, we tried to keep that status.

Even though we compile the libraries and have a patch in clang,
which enables an aarch64 Bitrig target, we do not enable that
in compile-time.  cmake used to take care of that in pots.

Therefore, enable the AArch64 target in the config files and
make the llvm binaries depend on those libraries.

No objections from pedro@.