fdtbus: enable driver attachment order
Previously we iterated over every fdt node and then looked for a
device that is compatible to it.  Unfortunately, this does not
work out well for us.

E.g. the i.MX6 clock driver needs to be run before any other driver.
To achieve that we iterate over every driver and then iterate over
every node and see if it's compatible.  Using this approach we can
control which driver attaches first, depending on the order in
the files.* file.