Pull in LLVM fix for PR20893.
Totally forget deallocated SDNodes in SDDbgInfo.

What would happen before that commit is that the SDDbgValues associated
with a deallocated SDNode would be marked Invalidated, but SDDbgInfo
would keep a map entry keyed by the SDNode pointer pointing to this list
of invalidated SDDbgNodes. As the memory gets reused, the list might get
wrongly associated with another new SDNode. As the SDDbgValues are
cloned when they are transfered, this can lead to an exponential number
of SDDbgValues being produced during DAGCombine like in
http://llvm.org/bugs/show_bug.cgi?id=20893

Note that the previous behavior wasn't really buggy as the invalidation
made sure that the SDDbgValues won't be used. This commit can be
considered a memory optimization and as such is really hard to validate
in a unit-test.