Bump datasizes for default and staff.
On typical desktops systems browsers already use enough
memory to hit the limit.  Also, we need to make a bit
of room for compiling base without hitting the limit.
LLVM seems to hitting the current limit quite easily.

Therefore bump the default datasize to 1024M, with a
maximum of 2048M.  Staff gains a default datasize
of 2048M and keeps the unlimited maximum.

ok jcv@ marco@ natano@ and more