Allow the disklabel to be read when an image is set via VNDIOCSET.
The call chain is as following: vndioctl -> disk_attach -> (workq) ->
disk_attach_callback -> disk_readlabel -> VOP_OPEN -> vndopen. In
vndopen the disklabel should be read for the coming DIOCGDINFO ioctl,
which doesn't happen because the device is still opened for the
VNDIOCSET ioctl.

This diff lets the vnd disk end up with a valid disklabel set after the
initial vnconfig call.

ok patrick@