Always allow abort tasks to be scheduled, even if the device is beeing detached, in order to prevent a deadlock situation.
This situation can occur if the thread detaching a device is sleeping,
waiting for all submitted transfers to finish, and the device's pipes
have not yet been aborted.  This can happen when a USB Ethernet device
is being detached while a userland program is doing an ioctl(2).

Abort tasks need to be able to run in such case since timed out transfers
rely on them to be properly completed.

ok deraadt@