New interrupt model, move away from IPLs.
This diff changes the interrupt model to something very similar to
what other modern unixes like Solaris, FreeBSD, DragonflyBSD and
linux. It also introduces critical sections.

Interrupts except clock and ipi are handled by interrupt threads,
when an interrupt fires, the only job for the small interrupt stub is
to schedule the corresponding interrupt thread, when the ithread gets
scheduled, it services the interrupt.

The kernel must be made preemptive, so that interrupt threads may
preempt the current running code.

In this model, you normally never block interrupts, you rely solely on
locks to protect the code, if the ithread preempts the running code,
and it tries to acquire a contested lock, it blocks (sleeps) and gets
resumed only when that lock is released.

This allows us to have lower interrupt latency, as instead of blocking
interrupts for a long section, you can fine grain that section in a
specific lock. Instead of raising to IPL_BIO and preventing softclock
from running for example, we can make softclock preempt the old
IPL_BIO and only block if there is an actual lock contention.

This was first implemented in Solaris 2 (Kleiman & Eykholt circa 93),
they demonstrated that this model, after properly locked, changed the
worst case latency from 1s to <1ms on a single-core 40mhz machine.

It also allows for us to properly fight livelocks in the future, as
the scheduler will make sure "everything runs at some point".

The kernel can now be synchronous with regard to locking, as you can
use the same lock to interlock against interrupts or other normal
kthreads.

You can only block on locks if you have a process context, so you
still need a way to block interrupts from normal (read: not ithread
interrupts), like clock and ipi. For that we introduce critical
sections, which blocks everything, in practice they're only used to
protect scheduler data and on mutexes, so they are very short.

In the future critical sections will also be the only thing that
prevents kernel preemption.

In order to prevent deadlocks, you must never be preempted while
holding a spinlock, so a critical section is implied there, this is
also akin to what every system does.

In this present state, kernel preemption has not been implemented, all
threads <IPL_CLOCK were moved to ithreads, there is no observable loss
of performance, it's been stable since the last half a year.

All spl calls <IPL_CLOCK were made no-ops, while >= IPL_CLOCK were
replaced by critical sections.

Most of the old assembly code has been rewritten in C, just because I
refuse to maintain unecessary asm blocks.

The next steps are:
    o Turn kernel lock into a rrwlock.
    o Enable kernel preemption, at this point all the interrupt
      interlocking will be done through kernel lock.
    o Decide on which subsystem to release first, having a wide subsytem lock.