Configure vnd via bio; ok pedro@
Integrate vnd into the bio framework. Adapt the management ioctls for
bio and rename them to reflect the changes in semantic: VNDIOCSET,
VNDIOCCLR and VNDIOCGET are now named BIOCVNDGET, BIOCVNDCLR and
BIOCVNDGET respectively.

This allows us to remove a workaround from vndopen, because the file
descriptor for the RAW_PARTITION is not open anymore when the device is
being configured.