add ml_filter and mq_filter functions to the mbuf list and queue apis.
this lets you run a filter function against each mbuf on a list or
queue. if the filter matches on an mbuf, it can return non-zero to
have ml_filter or mq_filter remove the mbuf and return it as part
of a chain of mbufs.

ok mpi@ claudio@ henning@ and s2k15 generally.