Don't compare against the pointer, it will always be true.
The blkno value should be increased every time we have a transfer
that has at least one byte. This check was broken, as it compared
the pointer to the len structure to zero, instead of actually
comparing the cmd length value. That always evaluates to true.

Fix this by comparing to the byte count value, which is set
a few lines earlier into the cmd len struct.