arm: scan the tree for interrupt controllers first
Even though you can usually assume that the nodes are
in the correct order in the FDT, we can't do that for
the interrupt controller.

Therefore we need to manually search for interrupt controllers
before scanning the rest of the tree.

We create a list of interrupt controllers which we will try
to get down to zero by attaching ever one of them. Still, we
need to take care of the interrupt controller hierarchy.

Therefore, we look for the root IC first, add it to the parent
list and then try to attach each parent's children.