arm: prototype of more MI intr_establish
Before this diff we had the issue that we
 * needed to somehow get the irq nr from the DT node
 * needed to call establish on the correct interrupt controller

This diff intends to solve theses issues. The idea is that you
only need to call an "MI" interrupt establish function and it
takes care of letting the correct interrupt controller now what
interrupt to establish.

This is implemented by adding a hook, so that interrupt controller
can register them as interrupt controllers. When the device driver
calls establish using the DT node, the code will look up the device's
interrupt controller and look for that one in the list.

Once found, it will call the IC's establish function and pass the
device node further down. The IC code will take care to properly
read and interpret the interrupt information and actually establish
the interrupt.