mkuboot: produce raw binaries if requested
mkuboot generates images bootable in u-boot.  These images
contain a header and the actual kernel object code.

This program was improved so it can embed kernel symbols
and the ramdisk into the generated image, like a bootloader
would do when booting from an actual ELF kernel.

Sometimes having a blob with ksyms+ramdisk _without_ the
image header for u-boot can be useful.  This change add
a flag 'r' which indicates that it should output the binary
in raw mode, without the u-boot headers.

"sounds fine then." drahn@