arm: fix build after LLVM update
Most of the changes are related to the change of default in which
as to use. Clang's new default is to use its integrated assembler.
Unfortunately there are a few incompatibilities, so some ASM files
need to be manually changed.

The libssl change is caused by the perl skript generating ASM
that clang doesn't think is valid, so we're gonna use the GNU as
for that. This needs to be revisited.

Also the directory structure in compiler-rt changed, so we need
to fix that up.