sdhc: add support for SD 3.0 specification
Version 3.0 adds higher frequencies. These are encoded a bit
differently in the capabilities registers.

Therefore we need to find out and store the spec version
and check for that when calculating the frequency.

Also, the divisor is calculated differently in version 3.0.

While there, add support to be able to pass the clkbase
from the attachment driver.

From NetBSD.

no objections pedro@