fdt: rework finding compatibles
Previously we assumed that node only have one compatible string.
This is wrong. A node can have multiple strings for propagating
what it is compatible to. This means we can't continue to look
for such nodes as we did. Instead we need to go through the
tree manually and check every node. Also we need to check every
string in the compatible property.