Fix x_e_getc() to not enter an endless loop.
The read loop in x_e_getc() failed to clear the got_sigwinch variable
and wound up in an endless loop. The bug (that has been introduced in
67223da12) can be triggered in various ways:

 * press ctrl-r, resize window
 * bind the version command to a keybinding and press it, resize window
 * press alt-1, resize window

The code in question was a special case for x_e_getc() that doesn't
redraw the prompt, when SIGWINCH arrives.  After careful inspection of
all the callers (x_e_getc(0)), I conclude that, in fact, all the callers
that used the non-redrawing variant did so in cases where redrawing
would have been beneficial. The diff at hand removes the 'redr'
parameter from x_e_getc(), such that that function always redraws the
prompt when SIGWINCH arrives, circumventing the bug described above.

reported by and ok pedro@